Move StatusBar.kt to a different package
